SUIT-15203 - handle then warn and exit lines
